implemented HTTP deleted by checking user's token and thier ids

+++ part4/controllers/blogs.js
@@ -20,8 +20,22 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blogId = request.params.id
+  const blog = await Blog.findById(blogId)
+
+  if (user.id.toString() === blog.user.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  else {
+    return response.status(401).json({ error: 'invalid user' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -52,7 +66,6 @@ blogsRouter.post('/', async (request, response) => {
   else {
     const body = request.body
     const token = request.token
-    console.log('middle', token)
     const decodedToken = jwt.verify(token, process.env.SECRET)
     if (!token || !decodedToken.id) {
       return response.status(401).json({ error: 'token missing or invalid' })
